progress on switching from int to uint to support allocating up to 4,000,000,000 items

+++ bintree/bininsert_nocomments.cpp
@@ -144,7 +144,7 @@ int bintree::insert(void* object, binaryTreeCompare lessThan)
   x->left = 0;
   x->right = 0;
 
-  memcpy(GETCLIENT(x), object, (size_t)tree->sizeOfClient);
+  memcpy(GETCLIENT(x), object, tree->sizeOfClient);
 
   BinInsert(tree, x);
